Refactored Frog Mating
Had to refactor the frog mating process to avoid the possiblity of breaking the engine loop when a single frog was born.  This caused the rest of the frogs that had not been iterated over yet to skip their chance at doing anything.
All frog mating chances are calculated and then processed at the end of the iteration and added to the main dictionary. This circumvents the errors I was getting with trying to add to the dictionary mid-iteration.